% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_BF_models.R
\name{available_models}
\alias{available_models}
\title{Get available models in rstanBF}
\usage{
available_models()
}
\value{
character vector of models
}
\description{
Get available models in rstanBF, as short names.
}
