% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_BF.R
\name{compute_BF_Stan}
\alias{compute_BF_Stan}
\title{Sample and compute BF for various models using Stan}
\usage{
compute_BF_Stan(data, model, hyperpriors, data_other = NULL,
  n.iter = 1000, n.burnin = 100, n.chains = 1, n.cores = 1,
  silent = FALSE)
}
\arguments{
\item{data}{a list containing \code{mtx}, \code{idx.ref}, \code{idx.quest}}

\item{model}{the model shortname (e.g. \code{'DirDir'}, \code{'DirFNorm'})}

\item{hyperpriors}{a list containing hyperparameter definitions}

\item{data_other}{a list containing additional data for $H_1$ and $H_2$ models (default: \code{NULL})}

\item{n.iter}{number of HMC iterations (default: 1000)}

\item{n.burnin}{number of HMC burn-in iterations (default: 100)}

\item{n.chains}{number of HMC chains (default: 1)}

\item{n.cores}{number of cores to use for HMC and bridgesampling (default: 1)}

\item{silent}{if TRUE, do not print any progress}
}
\value{
a \code{stanBF} object
}
\description{
Sample and compute BF for various models using Stan.
}
\details{
Two-sample model:
\itemize{
\item samples are stored as matrix \code{mtx}
\item reference items are indexed in \code{mtx} by \code{idx.ref}
\item questioned items are indexed in \code{mtx} by \code{idx.quest}
\item other items, non indexed, are \strong{discarded}.
}

Hypotheses:
\itemize{
\item $H_1$: samples in \code{idx.ref} and \code{idx.quest} come from the same source
\item $H_2$: samples in \code{idx.ref} and \code{idx.quest} come from the different sources
}

Return a \code{stanBF} object with these properties:
\itemize{
\item \code{model_name}
\item \code{stanmodel} (named list of Stan models)
\item \code{stanfit} (named list of stanfit objects)
\item \code{stanbridge} (named list of bridgesampler objects)
\item \code{BF} (a double)
}

For Dirichlet likelihoods, the returned object is a \code{stanBF_turn}, inheriting from \code{stanBF}.
These objects contain also:
\itemize{
\item \code{df_samples} (data.frame with posterior samples)
}

The object contains methods to plot and to extract samples.
}
