% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sirt_dirichlet.R
\name{digamma1}
\alias{digamma1}
\title{ML estimator of Dirichlet parameters}
\usage{
digamma1(x, h = 0.001)
}
\description{
From package \code{sirt}:
https://github.com/alexanderrobitzsch/sirt
}
\details{
GPL3 license
derivative of digamma function
}
