% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_dirichlet_MLE.R
\name{fun_Dirichlet_MLE_from_samples}
\alias{fun_Dirichlet_MLE_from_samples}
\title{Estimate MLE parameters from a dataframe of Dirichlet samples from different sources}
\usage{
fun_Dirichlet_MLE_from_samples(df_samples, col_source = source, ...)
}
\arguments{
\item{df_samples}{dataframe of Dirichlet samples}

\item{col_source}{the column name, unquoted, containing the source column}
}
\value{
a dataframe containing the Dirichlet parameter estimates for each source
}
\description{
Obtain estimates for each source.
}
