% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_dirichlet_MLE.R
\name{fun_Dirichlet_MLE_from_single_source}
\alias{fun_Dirichlet_MLE_from_single_source}
\title{Estimate Dirichlet parameter from a dataframe using ML or naive estimator}
\usage{
fun_Dirichlet_MLE_from_single_source(df, name_param = "theta",
  use = "MLE")
}
\arguments{
\item{df}{the dataframe containing only samples}

\item{name_param}{name of output matrix columns}

\item{use}{if 'MLE', use MLE estimator, else the sample mean (unbiased for alpha)}
}
\value{
a dataframe with the Dirichlet parameter named columns
}
\description{
Assume that data is ~ Dir(alpha).
}
\details{
This function returns a matrix containing the estimate for alpha.
}
