% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_BF_hyperpriors.R
\name{fun_estimate_DirDirGamma_hyperparameter}
\alias{fun_estimate_DirDirGamma_hyperparameter}
\title{Compute Dirichlet-DirichletGamma hyperparameters, using 'MLE' estimators}
\usage{
fun_estimate_DirDirGamma_hyperparameter(
  df_background,
  col_source = "source",
  eps = 1e-12,
  convcrit = 1e-08
)
}
\arguments{
\item{df_background}{dataframe with background data}

\item{col_source}{name of the source column (default: 'source')}

\item{eps}{convergence parameters for Dirichlet MLE (see: \code{\link[=dirichlet.mle]{dirichlet.mle()}})}

\item{convcrit}{convergence parameters for Dirichlet MLE (see: \code{\link[=dirichlet.mle]{dirichlet.mle()}})}
}
\value{
a numeric vector for the estimated Dirichlet hyperparameter
}
\description{
For the \code{'DirDirGamma'} model.
It computes an estimate using ML estimates of the parameters in each group.
}
\details{
Returns a list.
}
