% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_dirichlet_MLE.R
\name{fun_estimate_Dirichlet_from_samples}
\alias{fun_estimate_Dirichlet_from_samples}
\title{Estimate Dirichlet parameters for samples from multiple sources}
\usage{
fun_estimate_Dirichlet_from_samples(df, col_source = "source", ...)
}
\arguments{
\item{df}{dataframe of Dirichlet samples with a source column}

\item{col_source}{the column name containing the source column}

\item{...}{
  Arguments passed on to \code{\link[=fun_estimate_Dirichlet_from_single_source]{fun_estimate_Dirichlet_from_single_source}}
  \describe{
    \item{\code{name_param}}{name of output parameter (default: \code{'theta'})}
    \item{\code{use}}{if \code{'ML'}, use MLE estimator, else the sample estimators (unbiased for \eqn{\theta}, but not optimal)}
  }}
}
\value{
a tibble containing the Dirichlet parameter estimates for each source
}
\description{
Estimate MLE parameters from a dataframe of Dirichlet samples from different sources.
The sources must be known.
}
\details{
Suppose that rows in source i are \eqn{X ~ Dir(\theta_i)} iid.
This function estimates \eqn{\theta_i}.
}
