% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_BF_hyperpriors.R
\name{fun_estimate_Dirichlet_hyperparameter}
\alias{fun_estimate_Dirichlet_hyperparameter}
\title{Compute Dirichlet hyperparameters, using 'MLE' or 'naive'}
\usage{
fun_estimate_Dirichlet_hyperparameter(df_background, method)
}
\arguments{
\item{df_background}{dataframe with background data}

\item{method}{'MLE' or 'naive'}
}
\description{
Returns a numeric vector
}
