% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{is.stanBF}
\alias{is.stanBF}
\title{Check if an object is a stanBF object}
\usage{
is.stanBF(x)
}
\arguments{
\item{x}{an object}
}
\value{
TRUE if an object is a 'stanBF' object
}
\description{
Check if an object is a 'stanBF' object.
}
