% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{label_parse}
\alias{label_parse}
\title{ggplot2 labeller function for plotmath labels}
\usage{
label_parse(labels)
}
\arguments{
\item{labels}{character vector of labels to be parsed}
}
\value{
a list of \code{expressions}
}
\description{
ggplot2 labeller function for plotmath labels.
To be used in ggplot2 scale_*_ functions as the \code{label} argument.
}
\details{
It is useful when breaks are plotmath-ready character vectors.
This function correctly formats them in order to be parsed and displayed as expressions.
}
\examples{
df <- data.frame(
   var = c('x[1]', 'x[2]', 'x[3]', 'alpha'),
   value = c(1,2,3,4)
)

ggplot2(df) +
   geom_point(aes(x = var, y = value)) +
   scale_x_discrete(label = label_parse)


}
