% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{make_tbl_variable_range}
\alias{make_tbl_variable_range}
\title{Make a tibble with columns representing a range}
\usage{
make_tbl_variable_range(x.samples, text, ...)
}
\arguments{
\item{x.samples}{a matrix or data.frame which will be converted and re-named to a tibble}

\item{text}{the base name for the new column names}

\item{...}{additional value-name pairs which will be added as new columns ( \code{\link[tibble:add_column]{tibble::add_column()}})}
}
\value{
a tibble
}
\description{
Make a tibble with columns representing a range.
All column will have the same base name, and will follow the format of \code{'text[col_idx]'}, indexed from 1 to \code{ncol(x.samples)}.
}
\details{
It can be used to convert messy \code{rstan::extract} output from multidimensional variables, to a more manageable form.
}
