% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{named_vector_to_tibble}
\alias{named_vector_to_tibble}
\title{Convert a named vector to a tibble}
\usage{
named_vector_to_tibble(v)
}
\arguments{
\item{v}{a named vector}
}
\value{
a tibble
}
\description{
Convert a named vector to a tibble.
The names must be unique.
}
