% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_BF.R
\name{plot_posteriors.stanBF_turn}
\alias{plot_posteriors.stanBF_turn}
\title{Make boxplots for turn-point posteriors}
\usage{
\method{plot_posteriors}{stanBF_turn}(stanBF, variable = NULL,
  type = "boxplots")
}
\arguments{
\item{stanBF}{a supported \code{stanBF} object}

\item{variable}{\code{'theta'} or \code{'rho'} (normalized theta)}

\item{type}{type of plot (default: `'boxplots'``)}
}
\value{
a ggplot plot
}
\description{
Make boxplots for turn-point posteriors.
}
