% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_BF.R
\name{posterior_pred}
\alias{posterior_pred}
\title{Extract posterior predictive distributions}
\usage{
posterior_pred(stanBF, ...)
}
\arguments{
\item{stanBF}{a \code{stanBF} object}

\item{...}{additional parameters}
}
\description{
Extract posterior predictive distributions.
}
