% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{replace_names}
\alias{replace_names}
\title{Replace names using a pattern.}
\usage{
replace_names(x, pattern, replacement, ...)
}
\arguments{
\item{x}{a named object}

\item{pattern}{regex pattern}

\item{replacement}{replacement}

\item{...}{
  Arguments passed on to \code{\link[stringr:str_replace_all]{stringr::str_replace_all}}
  \describe{
    \item{\code{string}}{Input vector. Either a character vector, or something
coercible to one.}
  }}
}
\value{
x with replaced names.
}
\description{
Replace names using a pattern.
Calls \code{\link[stringr:str_replace_all]{stringr::str_replace_all()}}.
}
\examples{

fruits <- c(one = "one apple", two = "two pears", three = "three bananas")
replace_names(fruits, '^([a-z])', 'fruit_\\\\1')
replace_names(fruits, '^([a-z])', toupper)
}
\seealso{
\code{\link[stringr:str_replace_all]{stringr::str_replace_all()}}
}
