% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstanBF-package.R
\docType{package}
\name{rstanBF-package}
\alias{rstanBF-package}
\alias{rstanBF}
\title{The 'rstanBF' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (2018). RStan: the R interface to Stan. R package version 2.18.2. http://mc-stan.org
}
