% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_BF.R
\name{samples}
\alias{samples}
\title{Extract posterior samples from a stanBF object}
\usage{
samples(x, ...)
}
\arguments{
\item{...}{other arguments}

\item{stanBF}{a \code{stanBF} object}
}
\description{
Extract posterior samples from a stanBF object
}
