% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_BF_data.R
\name{stanBF_prepare_rsamplestudy_data}
\alias{stanBF_prepare_rsamplestudy_data}
\title{Convert the output of rsamplestudy to rstanBF data parameter}
\usage{
stanBF_prepare_rsamplestudy_data(list_pop, list_samples)
}
\arguments{
\item{list_pop}{the output of a \code{rsamplestudy}-generated population}

\item{list_samples}{the output of \code{\link[rsamplestudy:make_dataset_splits]{rsamplestudy::make_dataset_splits()}}}
}
\value{
a list containing \code{mtx}, \code{idx.ref}, \code{idx.quest}
}
\description{
Convert the output of rsamplestudy to rstanBF data parameter
}
