% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_BF.R
\name{samples.stanBF_turn}
\alias{samples.stanBF_turn}
\title{Exctract theta posterior samples for a turn-like object}
\usage{
\method{samples}{stanBF_turn}(stanBF)
}
\arguments{
\item{stanBF}{a \code{stanBF_turn} object}
}
\description{
\code{stanBF_turn} objects share the Dirichlet likelihood, with \eqn{\theta} as prior parameter.
}
\details{
This function extract posterior samples for \eqn{\theta}.
Also returns the normalized version of them (\eqn{\rho}).
}
