% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_BF_models.R
\name{available_models}
\alias{available_models}
\title{Get available models in rstanBF}
\usage{
available_models(verbose = FALSE)
}
\arguments{
\item{verbose}{get more information, in a data.frame form}
}
\value{
character vector of models, or a tibble
}
\description{
Get available models in rstanBF, as short names or longer form.
}
\details{
Optionally get more information in a tibble form
}
