% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_dirichlet_MLE.R
\name{fun_Dirichlet_MLE_from_single_source}
\alias{fun_Dirichlet_MLE_from_single_source}
\title{Estimate Dirichlet parameter from a dataframe using ML or naive estimator}
\usage{
fun_Dirichlet_MLE_from_single_source(df, name_param = "theta",
  use = "MLE")
}
\arguments{
\item{df}{the dataframe, must contain only the samples}

\item{name_param}{name of output parameter (default: \code{'theta'})}

\item{use}{if \code{'MLE'}, use MLE estimator, else the sample mean (unbiased for \eqn{\theta})}
}
\value{
a dataframe (tibble) with the Dirichlet parameter named columns
}
\description{
Assume that data is \eqn{X ~ Dir(\theta)} iid.
This function estimates \eqn{\theta}.
}
