% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_BF_hyperpriors.R
\name{fun_estimate_Dirichlet_hyperparameter}
\alias{fun_estimate_Dirichlet_hyperparameter}
\title{Compute Dirichlet hyperparameters, using 'MLE' or 'naive' estimators}
\usage{
fun_estimate_Dirichlet_hyperparameter(
  df_background,
  method,
  col_source = "source"
)
}
\arguments{
\item{df_background}{dataframe with background data}

\item{method}{\code{'ML'} or \code{'naive'} (see \code{\link[=fun_estimate_Dirichlet_from_samples]{fun_estimate_Dirichlet_from_samples()}})}

\item{col_source}{name of the source column (default: 'source')}
}
\value{
a numeric vector for the estimated Dirichlet hyperparameter
}
\description{
For the \code{'DirDir'} model.
It computes an estimate using ML estimates of the parameters in each group.
}
\details{
Returns a numeric vector.
}
