% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_BF_hyperpriors.R
\name{fun_estimate_Dirichlet_hyperparameter}
\alias{fun_estimate_Dirichlet_hyperparameter}
\title{Compute Dirichlet hyperparameters, using 'MLE' or 'naive' estimators}
\usage{
fun_estimate_Dirichlet_hyperparameter(df_background, method,
  col_source = "source")
}
\arguments{
\item{df_background}{dataframe with background data}

\item{method}{'ML' or 'naive'}

\item{col_source}{name of the source column (default: 'source')}
}
\description{
For the 'DirichletDirichlet' model.
It computes the best
}
\details{
Returns a numeric vector
}
