% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_BF.R
\name{plot_posteriors}
\alias{plot_posteriors}
\title{Plot posterior distributions of a stanBF object.}
\usage{
plot_posteriors(x, ...)
}
\arguments{
\item{x}{a \code{stanBF} object}

\item{...}{other arguments}
}
\description{
Plot posterior distributions of a stanBF object.
}
