% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_BF.R
\name{plot_posteriors.stanBF_turn}
\alias{plot_posteriors.stanBF_turn}
\title{Make boxplots for turn-point posteriors}
\usage{
\method{plot_posteriors}{stanBF_turn}(obj_turn, variable = NULL,
  type = "boxplots")
}
\arguments{
\item{obj_turn}{a supported \code{stanBF} object}

\item{variable}{'theta' or 'rho' (normalized theta)}
}
\value{
a ggplot plot
}
\description{
Make boxplots for turn-point posteriors.
}
