% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_BF.R
\name{print.stanBF}
\alias{print.stanBF}
\title{Print a \code{stanBF} object}
\usage{
\method{print}{stanBF}(stanBF, verbose = FALSE)
}
\arguments{
\item{stanBF}{a \code{stanBF} object}

\item{verbose}{print more details}
}
\description{
Print a \code{stanBF} object
}
