% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_BF.R
\name{prior_pred}
\alias{prior_pred}
\title{Extract prior predictive distributions}
\usage{
prior_pred(x, ...)
}
\arguments{
\item{x}{a \code{stanBF} object}
}
\description{
Extract prior predictive distributions
}
