% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_BF.R
\name{prior_pred}
\alias{prior_pred}
\title{Extract prior predictive distributions}
\usage{
prior_pred(stanBF, ...)
}
\arguments{
\item{stanBF}{a \code{stanBF} object}

\item{...}{additional parameters}
}
\description{
Extract prior predictive distributions
}
