% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_BF.R
\name{prior_pred.stanBF_turn}
\alias{prior_pred.stanBF_turn}
\title{Extract prior predictive distributions for turn-point posteriors.}
\usage{
prior_pred.stanBF_turn(obj_StanBF, ...)
}
\arguments{
\item{x}{a \code{stanBF_turn} object}
}
\value{
a list of data.frames
}
\description{
Extract prior predictive distributions for turn-point posteriors.
}
