% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_BF_samples.R
\name{samples}
\alias{samples}
\title{Extract posterior samples from a stanBF object}
\usage{
samples(stanBF, ...)
}
\arguments{
\item{stanBF}{a \code{stanBF} object}

\item{...}{other arguments}
}
\description{
Extract posterior samples from a stanBF object
}
