% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_BF_hyperpriors.R
\name{stan_BF_elicit_hyperpriors}
\alias{stan_BF_elicit_hyperpriors}
\title{Elicit hyperpriors to use for Stan BF}
\usage{
stan_BF_elicit_hyperpriors(df_background, model, mode_hyperparameter, ...)
}
\arguments{
\item{df_background}{the dataframe containing background data, as returned by \code{\link[rsamplestudy:make_dataset_splits]{rsamplestudy::make_dataset_splits()}}}

\item{model}{the model short name}

\item{mode_hyperparameter}{how the parameters are estimated}
}
\value{
a list of hyperprior parameters
}
\description{
Elicit hyperpriors to use for Stan BF
}
