% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_BF_hyperpriors.R
\name{stanBF_elicit_hyperpriors}
\alias{stanBF_elicit_hyperpriors}
\title{Elicit hyperpriors to use for Stan BF}
\usage{
stanBF_elicit_hyperpriors(
  df_background,
  model,
  mode_hyperparameter,
  col_source = "source",
  ...
)
}
\arguments{
\item{df_background}{the dataframe containing background data, with the source column}

\item{model}{the model short name}

\item{mode_hyperparameter}{how the parameters are estimated (can be \code{'ML'} or \code{'vague'})}

\item{col_source}{name of the source column (default: \code{'source'})}

\item{...}{arguments to hyperprior estimation methods}
}
\value{
a list of hyperprior parameters, as many as expected by \code{\link[=compute_BF_Stan]{compute_BF_Stan()}}
}
\description{
Elicit hyperpriors to use for Stan BF
}
