% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_BF.R
\name{samples.stanBF_turn}
\alias{samples.stanBF_turn}
\alias{prior_pred.stanBF_turn}
\title{Extract theta posterior samples for a turn-like object}
\usage{
\method{samples}{stanBF_turn}(stanBF)

\method{prior_pred}{stanBF_turn}(obj_StanBF, ...)
}
\arguments{
\item{stanBF}{a \code{stanBF_turn} object}

\item{x}{a \code{stanBF_turn} object}
}
\value{
a tibble containing prior predictions across hypotheses and sources
}
\description{
\code{stanBF_turn} objects share the Dirichlet likelihood, with \eqn{\theta} as prior parameter.

Extract prior predictive distributions for turn-point posteriors.
}
\details{
This function extract posterior samples for \eqn{\theta}.
Also returns the normalized version of them (\eqn{\rho}).
}
